refactor(evals): move evaluator delete button
