+++ web/src/components/deleteButton.tsx
@@ -1,5 +1,4 @@
 import { useMemo, useState } from "react";
-import Link from "next/link";
 import { useRouter } from "next/router";
 import {
   Popover,
@@ -7,7 +6,7 @@ import {
   PopoverTrigger,
 } from "@/src/components/ui/popover";
 import { Button, type ButtonProps } from "@/src/components/ui/button";
-import { ExternalLinkIcon, LockIcon, TrashIcon } from "lucide-react";
+import { LockIcon, TrashIcon } from "lucide-react";
 import { useHasProjectAccess } from "@/src/features/rbac/utils/checkProjectAccess";
 import { type ProjectScope } from "@/src/features/rbac/constants/projectAccessRights";
 import { api } from "@/src/utils/api";
@@ -15,7 +14,6 @@ import { usePostHogClientCapture } from "@/src/features/posthog-analytics/usePos
 import { Input } from "@/src/components/ui/input";
 import { Label } from "@/src/components/ui/label";
 import { showSuccessToast } from "@/src/features/notifications/showSuccessToast";
-import { showErrorToast } from "@/src/features/notifications/showErrorToast";
 import { useHasEntitlement } from "@/src/features/entitlements/hooks";
 
 export type DeleteButtonProps = {
@@ -474,134 +472,6 @@ export function DeleteEvaluationModelButton(
       deleteConfirmation="delete"
       executeDeleteMutation={executeDeleteMutation}
       isDeleteMutationLoading={isPending}
-    />
-  );
-}
-
-export function DeleteEvalTemplateButton(
-  props: DeleteButtonProps & {
-    // usage count from already-loaded table data; lets the blocked state
-    // render instantly while the usage query is in flight
-    initialUsageCount?: number;
-    // the template name; doubles as the type-to-confirm string
-    deleteConfirmation: string;
-  },
-) {
-  const utils = api.useUtils();
-  const {
-    itemId,
-    projectId,
-    scope = "evalTemplate:CUD",
-    invalidateFunc = () => utils.evals.invalidate(),
-    initialUsageCount,
-    deleteConfirmation: templateName,
-  } = props;
-
-  const templateMutation = api.evals.deleteEvalTemplate.useMutation({
-    onSuccess: () => {
-      showSuccessToast({
-        title: "Evaluator deleted",
-        description: `Evaluator "${templateName}" was deleted.`,
-      });
-      utils.evals.invalidate();
-    },
-    onError: (error) =>
-      showErrorToast("Failed to delete evaluator", error.message),
-  });
-
-  // only fetch usage once the user actually opens the delete popover; once
-  // the delete starts the query must go inactive, otherwise the post-delete
-  // cache invalidation refetches it and surfaces a NOT_FOUND. Gating on
-  // isPending (not just isSuccess) matters: the mutation's onSuccess — and
-  // its invalidate — runs before the success state is dispatched, so the
-  // query must already be disabled by then.
-  const [hasOpenedPopover, setHasOpenedPopover] = useState(false);
-  const usage = api.evals.evalTemplateUsage.useQuery(
-    { projectId, evalTemplateId: itemId },
-    {
-      enabled:
-        hasOpenedPopover &&
-        !templateMutation.isPending &&
-        !templateMutation.isSuccess,
-    },
-  );
-  const referencingEvaluators = usage.data;
-  const usageCount = referencingEvaluators?.length ?? initialUsageCount ?? 0;
-
-  const executeDeleteMutation = async (onSuccess: () => void) => {
-    try {
-      await templateMutation.mutateAsync({
-        evalTemplateId: itemId,
-        projectId,
-      });
-      onSuccess();
-    } catch {
-      // surfaced via the mutation's onError toast (e.g. evaluator still in use)
-    }
-  };
-
-  const renderEvaluatorLink = (evaluator: {
-    id: string;
-    scoreName: string;
-  }) => (
-    <Link
-      href={`/project/${projectId}/evals?peek=${evaluator.id}`}
-      target="_blank"
-      rel="noopener noreferrer"
-      className="inline-flex items-center gap-1 text-blue-600 hover:underline"
-    >
-      {evaluator.scoreName}
-      <ExternalLinkIcon className="h-3 w-3" />
-    </Link>
-  );
-
-  const deleteBlocker =
-    usageCount > 0 ? (
-      <>
-        <h2 className="text-md mb-3 font-semibold">Cannot delete</h2>
-        <p className="mb-3 max-w-72 text-sm">
-          This evaluator is used by {usageCount} running evaluator
-          {usageCount === 1 ? "" : "s"}. Delete those running evaluators first.
-        </p>
-        {referencingEvaluators && referencingEvaluators.length > 0 ? (
-          referencingEvaluators.length === 1 ? (
-            <div className="max-w-72 text-sm">
-              {renderEvaluatorLink(referencingEvaluators[0])}
-            </div>
-          ) : (
-            <ul className="max-h-40 max-w-72 list-inside list-disc overflow-y-auto text-sm">
-              {referencingEvaluators.map((evaluator) => (
-                <li key={evaluator.id}>{renderEvaluatorLink(evaluator)}</li>
-              ))}
-            </ul>
-          )
-        ) : null}
-      </>
-    ) : undefined;
-
-  return (
-    <DeleteButton
-      {...props}
-      scope={scope}
-      invalidateFunc={invalidateFunc}
-      captureDeleteOpen={(capture, isTableAction) =>
-        capture("eval_templates:delete_form_open", {
-          source: isTableAction ? "table-single-row" : "template",
-        })
-      }
-      captureDeleteSuccess={(capture, isTableAction) =>
-        capture("eval_templates:delete_template_button_click", {
-          source: isTableAction ? "table-single-row" : "template",
-        })
-      }
-      entityToDeleteName="evaluator"
-      customDeletePrompt="This action cannot be undone. It permanently deletes all versions of this evaluator. Scores already produced by it will not be deleted."
-      executeDeleteMutation={executeDeleteMutation}
-      isDeleteMutationLoading={templateMutation.isPending}
-      deleteBlocker={deleteBlocker}
-      onPopoverOpenChange={(open) => {
-        if (open) setHasOpenedPopover(true);
-      }}
     />
   );
 }
