+++ web/src/features/evals/components/eval-templates-table.tsx
@@ -31,7 +31,7 @@ import {
   DropdownMenuLabel,
   DropdownMenuTrigger,
 } from "@/src/components/ui/dropdown-menu";
-import { DeleteEvalTemplateButton } from "@/src/components/deleteButton";
+import { DeleteEvalTemplateButton } from "@/src/features/evals/components/delete-eval-template-button";
 import { EvalTemplateForm } from "@/src/features/evals/components/template-form";
 import { showSuccessToast } from "@/src/features/notifications/showSuccessToast";
 import { EvalReferencedEvaluators } from "@/src/features/evals/types";
